+++ dhondtxai/__init__.py
@@ -2,7 +2,7 @@
 from .dhondt_xai import DhondtExplanation, DhondtValues, DhondtXAI
 from .plot_parliament import plot_parliament, plot_signed_parliament
 
-__version__ = "0.9.2"
+__version__ = "0.9.3"
 Explainer = DhondtXAI
 
 __all__ = [